Refactor business hours utility
